split "installation" logic from "bubblewrap" logic
